+++ addons/yat/src/scenes/game_terminal/components/quick_commands_context/QuickCommandsContext.cs
@@ -46,9 +46,9 @@ public partial class QuickCommandsContext : ContextSubmenu
             return;
         }
 
-        _ = _terminalSwitcher.CurrentTerminal.CommandManager.RunAsync(
+        _ = _terminalSwitcher.CurrentTerminal.CommandManager.Run(
             Text.SanitizeText(command),
             _terminalSwitcher.CurrentTerminal
-        ).ConfigureAwait(false).GetAwaiter().GetResult();
+        );
     }
 }